refactored `binding` to `bindings`
