+++ ceres-bindings/src/main/kotlin/dev/teogor/ceres/bindings/MaterialBinding.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.teogor.ceres.binding
+package dev.teogor.ceres.bindings
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
